Don't let `main` return `Result`
